# PERF: merge on monotonic keys